Final iteration of package spltting mock server packages
