+++ apps/mock-oauth2-server/.env
@@ -1,0 +1,9 @@
+# Personal User Credentials
+EMAIL=nkduy2010@gmail.com
+FAMILY_NAME=Nguyen
+GIVEN_NAME=Duy
+
+# Admin User Credentials (optional)
+ADMIN_EMAIL=nkduy2010+admin@gmail.com
+ADMIN_FAMILY_NAME=Nguyen
+ADMIN_GIVEN_NAME=Duy
