+++ apps/mock-oauth2-server/tsconfig.json
@@ -1,0 +1,11 @@
+{
+	"extends": "@cellix/typescript-config/node.json",
+	"compilerOptions": {
+		"rootDir": ".",
+		"outDir": "dist"
+	},
+	"include": ["src/**/*.ts"],
+	"references": [
+		{ "path": "../../packages/cellix/mock-oauth2-server-seedwork" }
+	]
+}
