+++ apps/mock-payment-server/.env
@@ -1,0 +1,1 @@
+# Local environment variables
