+++ apps/mock-payment-server/.gitignore
@@ -1,0 +1,5 @@
+node_modules/
+dist/
+.env
+tsconfig.tsbuildinfo
+.turbo
