+++ apps/mock-payment-server/package.json
@@ -1,0 +1,28 @@
+{
+	"name": "@app/mock-payment-server",
+	"version": "1.0.0",
+	"main": "dist/src/index.js",
+	"types": "dist/src/index.d.ts",
+	"type": "module",
+	"license": "MIT",
+	"description": "Local payment mock server for dev/testing",
+	"scripts": {
+		"prebuild": "biome lint",
+		"build": "tsc --build",
+		"clean": "rimraf dist node_modules tsconfig.tsbuildinfo && tsc --build --clean",
+		"lint": "biome lint",
+		"format": "biome format --write",
+		"start": "node dist/src/index.js",
+		"dev": "tsx watch src/index.ts"
+	},
+	"dependencies": {
+		"@cellix/mock-payment-server-seedwork": "workspace:*",
+		"dotenv": "^16.4.5"
+	},
+	"devDependencies": {
+		"@cellix/typescript-config": "workspace:*",
+		"rimraf": "^6.0.1",
+		"tsx": "^4.20.3",
+		"typescript": "^5.8.3"
+	}
+}
