+++ apps/mock-payment-server/turbo.json
@@ -1,0 +1,4 @@
+{
+	"extends": ["//"],
+	"tags": ["backend"]
+}
