+++ packages/cellix/mock-oauth2-server-seedwork/.gitignore
@@ -1,0 +1,5 @@
+node_modules/
+dist/
+.env
+tsconfig.tsbuildinfo
+.turbo
