+++ packages/cellix/mock-oauth2-server-seedwork/package.json
@@ -1,0 +1,41 @@
+{
+    "name": "@cellix/mock-oauth2-server-seedwork",
+    "version": "1.0.0",
+    "private": true,
+    "license": "MIT",
+    "description": "Mock OAuth2/OIDC server seedwork for CellixJS monorepo.",
+    "type": "module",
+    "main": "dist/src/index.js",
+    "types": "dist/src/index.d.ts",
+    "exports": {
+        ".": {
+            "types": "./dist/src/index.d.ts",
+            "import": "./dist/src/index.js"
+        },
+        "./setup-environment": {
+            "types": "./dist/src/setup-environment.d.ts",
+            "import": "./dist/src/setup-environment.js"
+        }
+    },
+    "scripts": {
+        "prebuild": "biome lint",
+        "build": "tsc --build",
+        "clean": "rimraf dist"
+    },
+    "peerDependencies": {
+		"express": "^4.0.0",
+		"jose": "^5.0.0"
+	},
+	"dependencies": {
+		"dotenv": "^16.4.5"
+	},
+    "devDependencies": {
+        "@cellix/typescript-config": "workspace:*",
+        "@types/express": "^4.17.23",
+        "dotenv": "^16.4.5",
+        "express": "^4.22.1",
+        "jose": "^5.9.6",
+        "rimraf": "^6.0.1",
+        "typescript": "^5.8.3"
+    }
+}
