+++ packages/cellix/mock-oauth2-server-seedwork/src/setup-environment.ts
@@ -1,0 +1,8 @@
+import dotenv from 'dotenv';
+
+export const setupEnvironment = () => {
+	console.log('Setting up environment variables');
+	dotenv.config();
+	dotenv.config({ path: `.env.local`, override: true });
+	console.log('Environment variables set up');
+};
