+++ packages/cellix/mock-oauth2-server-seedwork/tsconfig.json
@@ -4,5 +4,6 @@
 		"rootDir": ".",
 		"outDir": "dist"
 	},
-	"include": ["src/**/*.ts"]
+	"include": ["src/**/*.ts"],
+	"references": []
 }
