--- packages/cellix/mock-oauth2-server/.env
@@ -1,9 +1,0 @@
-# Personal User Credentials
-Email=nkduy2010@gmail.com
-Family_Name=Nguyen
-Given_Name=Duy
-
-# Admin User Credentials (optional)
-Admin_Email=nkduy2010+admin@gmail.com
-Admin_Family_Name=Nguyen
-Admin_Given_Name=Duy
