--- packages/cellix/mock-oauth2-server/.gitignore
@@ -1,7 +1,0 @@
-# .gitignore template - https://github.com/github/gitignore/blob/main/Node.gitignore
-/node_modules
-.DS_Store
-__*
-dist
-tsconfig.tsbuildinfo
-.turbo
