--- packages/cellix/mock-oauth2-server/package.json
@@ -1,31 +1,0 @@
-{
-    "name": "@cellix/mock-oauth2-server",
-    "version": "1.0.0",
-    "author": "",
-    "license": "MIT",
-    "description": "Local OAuth2/OIDC mock server for dev/testing",
-    "type": "module",
-    "main": "dist/src/index.js",
-    "types": "dist/src/index.d.ts",
-    "scripts": {
-        "build": "tsc --build",
-        "clean": "rimraf dist node_modules tsconfig.tsbuildinfo && tsc --build --clean",
-        "lint": "biome lint",
-        "format": "biome format --write",
-        "start": "node dist/src/index.js",
-        "dev": "tsx watch src/index.ts",
-        "prebuild": "biome lint"
-    },
-    "dependencies": {
-        "dotenv": "^16.4.5",
-        "express": "^4.22.1",
-        "jose": "^5.9.6"
-    },
-    "devDependencies": {
-        "@cellix/typescript-config": "workspace:*",
-        "@types/express": "^4.17.23",
-        "rimraf": "^6.0.1",
-        "tsx": "^4.20.3",
-        "typescript": "^5.8.3"
-    }
-}
