--- packages/cellix/mock-oauth2-server/src/setup-environment.ts
@@ -1,8 +1,0 @@
-import dotenv from 'dotenv';
-
-export const setupEnvironment = () => {
-    console.log('Setting up environment variables');
-    dotenv.config();
-    dotenv.config({ path: `.env.local`, override: true });
-    console.log('Environment variables set up');
-}
