--- packages/cellix/mock-oauth2-server/turbo.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["//"],
-  "tags": ["backend"]
-}
