+++ packages/cellix/mock-payment-server-seedwork/package.json
@@ -1,0 +1,38 @@
+{
+	"name": "@cellix/mock-payment-server-seedwork",
+	"version": "1.0.0",
+	"main": "dist/src/index.js",
+	"types": "dist/src/index.d.ts",
+	"type": "module",
+	"license": "MIT",
+	"private": true,
+	"description": "Mock payment server seedwork for CellixJS monorepo.",
+	"exports": {
+		".": {
+			"types": "./dist/src/index.d.ts",
+			"import": "./dist/src/index.js"
+		}
+	},
+	"scripts": {
+		"prebuild": "biome lint",
+		"build": "tsc --build && node dist/src/copy-assets.js",
+		"clean": "rimraf dist"
+	},
+	"peerDependencies": {
+		"@cellix/payment-service": "workspace:*",
+		"express": "^4.0.0",
+		"jose": "^5.0.0",
+		"jsonwebtoken": "^9.0.0"
+	},
+	"devDependencies": {
+		"@cellix/typescript-config": "workspace:*",
+		"@cellix/payment-service": "workspace:*",
+		"@types/express": "^4.17.21",
+		"@types/jsonwebtoken": "^9.0.10",
+		"express": "^4.22.0",
+		"jose": "^5.10.0",
+		"jsonwebtoken": "^9.0.3",
+		"rimraf": "^6.0.1",
+		"typescript": "^5.0.0"
+	}
+}
