+++ packages/cellix/mock-payment-server-seedwork/src/copy-assets.ts
