--- packages/cellix/mock-payment-server/package.json
@@ -1,28 +1,0 @@
-{
-	"name": "@cellix/mock-payment-server",
-	"version": "1.0.0",
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
-	"type": "module",
-	"license": "MIT",
-	"dependencies": {
-		"@cellix/payment-service": "workspace:*",
-		"express": "^4.22.0",
-		"jose": "^5.10.0",
-		"jsonwebtoken": "^9.0.3"
-	},
-	"devDependencies": {
-		"@cellix/typescript-config": "workspace:*",
-		"@types/express": "^4.17.21",
-		"@types/jsonwebtoken": "^9.0.10",
-		"ts-node": "^10.9.2",
-		"ts-node-dev": "^2.0.0",
-		"tsc-watch": "^7.1.1",
-		"typescript": "^5.0.0"
-	},
-	"scripts": {
-		"start": "node dist/src/index.js",
-		"build": "tsc --build && node dist/src/copy-assets.js",
-		"clean": "rimraf node_modules package-lock.json dist"
-	}
-}
