--- packages/cellix/mock-payment-server/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@cellix/typescript-config/node.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "."
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules", "dist"]
-}
