--- packages/cellix/mock-payment-server/turbo.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["//"],
-  "tags": ["backend"]
-}
